Relocate Spec\ZonedDateTime @internal helpers to Spec\Internal\
Two `@internal public static` methods (normalizeTimezoneId,
wallSecToEpochSec) sat on the public Spec\ZonedDateTime surface,
contradicting the BC promise on day one. Move them — together with
their private helpers normalizeTimezoneIdUncached,
wallSecToEpochSecStartOfDay, and safeGetTransitions — to a new
Spec\Internal\TimeZoneHelper. The third (createFromEpochParts) needs
access to ZonedDateTime's private trueEpochSec/trueSubNs fields and
naturally belongs on the class itself; drop its @internal tag and
rewrite the docblock to read as a public factory.

Update internal callers in Spec\ZonedDateTime and external callers in
Spec\Duration, Spec\Instant, Spec\PlainDate, Spec\PlainDateTime to
route through TimeZoneHelper. Drops the stranded @internal docblock
that was sandwiched between two unrelated docblocks above
wallSecToEpochSecStartOfDay.

Closes #16.
